$dom JS. Load workers scripts from script tags, load remote or local scripts.
I like jquery and underscore but I really like knowing how things work and re-creating things that I see from scratch. I write most of the stuff in my repos on a cromebook that is rather limited as far as a development platform goes to this library has some functions aimed at getting around security restrictions on the chromebook. Mainly the ability to work with Webworkers. and loading scripts locally.